corrección de errores menores y nuevos de  metodos

+++ app/src/main/java/com/innovati/felipehernandez/invenenvios/pojos/ArticulosPedido.java
@@ -4,17 +4,17 @@ public class ArticulosPedido {
     String idArticulo;
     String nombre;
     String presentacion;
-    double cantidad;
-    double precio;
-    double subTotal;
-    double iva;
-    double total;
+    float cantidad;
+    float precio;
+    float subTotal;
+    float iva;
+    float total;
     boolean status;
 
     public ArticulosPedido() {
     }
 
-    public ArticulosPedido(String idArticulo, String nombre, String presentacion, double cantidad, double precio, double subTotal, double iva, double total, boolean status) {
+    public ArticulosPedido(String idArticulo, String nombre, String presentacion, float cantidad, float precio, float subTotal, float iva, float total, boolean status) {
         this.idArticulo = idArticulo;
         this.nombre = nombre;
         this.presentacion = presentacion;
@@ -58,43 +58,43 @@ public class ArticulosPedido {
         this.presentacion = presentacion;
     }
 
-    public double getCantidad() {
+    public float getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(double cantidad) {
+    public void setCantidad(float cantidad) {
         this.cantidad = cantidad;
     }
 
-    public double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(float precio) {
         this.precio = precio;
     }
 
-    public double getSubTotal() {
+    public float getSubTotal() {
         return subTotal;
     }
 
-    public void setSubTotal(double subTotal) {
+    public void setSubTotal(float subTotal) {
         this.subTotal = subTotal;
     }
 
-    public double getIva() {
+    public float getIva() {
         return iva;
     }
 
-    public void setIva(double iva) {
+    public void setIva(float iva) {
         this.iva = iva;
     }
 
-    public double getTotal() {
+    public float getTotal() {
         return total;
     }
 
-    public void setTotal(double total) {
+    public void setTotal(float total) {
         this.total = total;
     }
 }
